+++ rollup.config.js
@@ -10,10 +10,12 @@ export default {
     {
       file: pkg.main,
       format: 'cjs',
+      exports: 'named',
     },
     {
       file: pkg.module,
       format: 'es',
+      exports: 'named',
     },
   ],
   plugins: [