+++ Changelog.md
@@ -2,6 +2,11 @@
 
 ## master
 
+## 0.0.4
+
+* Expose StubbornFetchError export
+* Add rollup watch
+
 ## 0.0.3
 
 * Update module output